refactor : 불필요한 클래스 삭제, EmailSender 모듈 변경
