+++ likelion-infrastructure/src/main/java/likelion/univ/email/sender/EmailSender.java
@@ -1,0 +1,6 @@
+package likelion.univ.email.sender;
+
+public interface EmailSender {
+
+    void send(EmailContent emailContent);
+}
